Refactor file strcuture for modules and test files accordingly
