--- modi/module/__init__.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Module package for pyMODI."""
-
-__author__ = """Jinsung Ha"""
-__email__ = "jinsung@luxrobo.com"
-__version__ = "0.7.0"
-
-__all__ = [
-    "env",
-    "gyro",
-    "mic",
-    "button",
-    "dial",
-    "ultrasonic",
-    "ir",
-    "display",
-    "motor",
-    "led",
-    "speaker",
-]
